moved files around, made changes to config
