+++ server/app.js
@@ -4,17 +4,16 @@ const morgan = require('morgan');
 const toDoRoutes = require('./routes/toDoRoutes');
 const authRoutes = require('./routes/authRoutes');
 const { auth } = require('express-openid-connect');
-const { db } = require('./db');
 require('dotenv').config();
 
 //express app
 const app = express();
-const PORT = process.env.PORT || 4000;
 
 //middleware
 app.use(express.json());
 app.use(morgan('dev'));
 app.use(express.urlencoded({ extended: true }));
+app.use(express.static('build'));
 
 const { AUTH0_SECRET, AUTH0_BASE_URL, AUTH0_CLIENT_ID, AUTH0_ISSUER_BASE_URL } =
   process.env;
@@ -38,7 +37,7 @@ app.use(auth(config));
 
 // Configure CORS for React frontend
 app.use(cors({
-  origin: 'http://localhost:4000',
+  origin: 'http://localhost:3000',
   methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
   credentials: true,
 }));
@@ -46,7 +45,7 @@ app.use(cors({
 //Home page - different response depending on if user is logged in or not
 app.get('/', (req, res) => {
   if (req.oidc.user) {
-    res.json({ response: `Hello, ${req.oidc.user.given_name}` }); //if logged in
+    res.json({ response: `He llo, ${req.oidc.user.given_name}` }); //if logged in
   } else {
     res.json({ response: 'Hello there!' }); //no login present
   }
@@ -55,9 +54,5 @@ app.get('/', (req, res) => {
 app.use('/auth', authRoutes); // login and logout
 app.use('/todos', toDoRoutes); //todo routes
 
-app.listen(PORT, () => {
-  db.sync();
-  console.log(`Server running on port: http://localhost:${PORT}`);
-});
 
 module.exports = app;
